requirements.txt
pytorch-lightning
torch
torchvision
numpy
Pillow
jsonargparse[signatures]
torchmetrics